-Added TTMessages, &viewmsg and other stuff so that the developers can send messages to specific players. (Issue Not Listed) -Fixed bug in irc client (Issue N/A) -Finished the Fun part of the GUI (Issue #14) -Updated &commands and &help (Issue N/A) -Fixed bug in Outfits.vb, calling GetOutfits returned an array with an invalid element at the end (Issue N/A)

+++ TibiaTekBot2/BattleList.vb
@@ -518,7 +518,7 @@ Public Module BattleListModule
             End Set
         End Property
 
-        Public Property OutfitID() As UShort
+        Public Property OutfitID() As Integer
             Get
                 Try
                     Dim Result As Integer = 0
@@ -529,7 +529,7 @@ Public Module BattleListModule
                     End
                 End Try
             End Get
-            Set(ByVal NewOutfit As UShort)
+            Set(ByVal NewOutfit As Integer)
                 Try
                     Core.WriteMemory(Consts.ptrBattleListBegin + (IndexPosition * Consts.BLDist) + Consts.BLOutfitOffset, NewOutfit, 2)
                 Catch Ex As Exception
